Updated the use case according to the review
Settled on the term 'honor' and removed any alternative terminology from the use case. Rewritten some aspects of the use case to distinct it from a similar 'like' feature in the photo layer of a map. Added a statement about the effects of the honors on a map on the order of results in the map search.