squash-commit
feat: add DebouncedSimpleFormInput

refactor: move SearchInput to Form/Input in storybook

feat: Area of Plantings Form

refactor: remove & rename planting attributes

remove `scale_x`, `scale_y` attributes.
A planting's scale should not be changing.
Instead the editor element's `size_x` and `size_y`
change to enable a scaling.

This was done to distinguish it from a plant's
height, that might be added later.
As a three-dimensional concept it can not be
represented on the editor map directly.

refactor: transformer to common api

fix: relations for area of plantings